entry point install wip
